Bunch of changes
created 2 divs for each batch of HTML content to isolate the content i
wanted to alter.
added link to bottom of fact list to make it easier to toggle between
hiding/unhiding the list of facts for each park.
wrote code to toggle the list of facts on and off